cedar-go/types: split up each type into its own file for legibility's sake
