+++ types/long_test.go
@@ -1,0 +1,36 @@
+package types_test
+
+import (
+	"testing"
+
+	"github.com/cedar-policy/cedar-go/internal/testutil"
+	"github.com/cedar-policy/cedar-go/types"
+)
+
+func TestLong(t *testing.T) {
+	t.Parallel()
+
+	t.Run("Equal", func(t *testing.T) {
+		t.Parallel()
+		one := types.Long(1)
+		one2 := types.Long(1)
+		zero := types.Long(0)
+		f := types.Boolean(false)
+		testutil.FatalIf(t, !one.Equal(one), "%v not Equal to %v", one, one)
+		testutil.FatalIf(t, !one.Equal(one2), "%v not Equal to %v", one, one2)
+		testutil.FatalIf(t, one.Equal(zero), "%v Equal to %v", one, zero)
+		testutil.FatalIf(t, zero.Equal(one), "%v Equal to %v", zero, one)
+		testutil.FatalIf(t, zero.Equal(f), "%v Equal to %v", zero, f)
+	})
+
+	t.Run("string", func(t *testing.T) {
+		t.Parallel()
+		AssertValueString(t, types.Long(1), "1")
+	})
+
+	t.Run("TypeName", func(t *testing.T) {
+		t.Parallel()
+		tn := types.Long(1).TypeName()
+		testutil.Equals(t, tn, "long")
+	})
+}
